refactor: move test clients to a client directory
Added a new 'client' directory inside unitTest/utils to store all
clients
