+++ backend/kernelCI_app/unitTests/utils/client/hardwareClient.py
@@ -1,6 +1,6 @@
 import requests
 from django.urls import reverse
-from kernelCI_app.unitTests.utils.baseClient import BaseClient
+from kernelCI_app.unitTests.utils.client.baseClient import BaseClient
 from kernelCI_app.typeModels.hardwareDetails import HardwareDetailsPostBody
 import json
