+++ backend/kernelCI_app/unitTests/utils/client/issueClient.py
@@ -1,7 +1,7 @@
 import requests
 from django.urls import reverse
 import json
-from kernelCI_app.unitTests.utils.baseClient import BaseClient
+from kernelCI_app.unitTests.utils.client.baseClient import BaseClient
 
 
 class IssueClient(BaseClient):
